Created error pages.

+++ blog/posts/routes.py
@@ -1,4 +1,4 @@
-from flask import Blueprint, render_template, redirect, url_for, request
+from flask import Blueprint, render_template, redirect, url_for, request, flash, abort
 from flask_login import login_required, current_user
 from blog.posts.forms import CreateOrUpdatePostForm
 from blog.comments.forms import CommentForm
@@ -23,7 +23,7 @@ def create_post():
         # Saving Post to database.
         new_post.save()
         return redirect(url_for("main.home"))
-    return render_template("posts/create_or_update_post.html", form=create_post_form, legend="New Post")
+    return render_template("posts/create_or_update_post.html", form=create_post_form, legend="New Post", title="New Post")
 
 # Post update route.
 @posts.route("/post/<post_id>/update", methods=["GET", "POST"])
@@ -31,16 +31,20 @@ def create_post():
 def update_post(post_id):
     update_form = CreateOrUpdatePostForm()
     post = Post.objects.get_or_404(id=post_id)
+    # Forbids other users to update posts.
+    if post.author != current_user:
+        abort(403)
     if update_form.validate_on_submit():
         post.title = update_form.title.data
         post.content = update_form.content.data
         post.save()
+        flash("Your post has been updated.", "success")
         return redirect(url_for("posts.comment_post", post_id=post_id))
     # Populating form.
     elif request.method == "GET":
         update_form.title.data = post.title
         update_form.content.data = post.content
-    return render_template("posts/create_or_update_post.html", form=update_form, legend="Update Post")
+    return render_template("posts/create_or_update_post.html", form=update_form, legend="Update Post", title="Update Post")
 
 # Post delete route.
 @posts.route("/post/<post_id>/delete", methods=["GET"])
@@ -48,9 +52,13 @@ def update_post(post_id):
 def delete_post(post_id):
     # Deleting post and all related comments to it.
     post = Post.objects.get_or_404(id=post_id)
+    # Forbids other users to delete posts.
+    if post.author != current_user:
+        abort(403)
     comments = Comment.objects(post=post)
     post.delete()
     comments.delete()
+    flash("Your post has been deleted.", "danger")
     return redirect(url_for("main.home"))
 
 # Comment post route.
@@ -70,4 +78,4 @@ def comment_post(post_id):
         )
         new_comment.save()
         return redirect(url_for("posts.comment_post", post_id=post_id))
-    return render_template("posts/comment_post.html", post=post, form=comment_form, comments=comments)
+    return render_template("posts/comment_post.html", post=post, form=comment_form, comments=comments, title="Post")
